requirements.txt
async-timeout==5.0.1
exceptiongroup==1.3.1
iniconfig==2.1.0
packaging==26.0
paho-mqtt==2.1.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.2.1
redis==7.0.1
tomli==2.4.0
typing_extensions==4.15.0
